// apps/web/src/api-client/model/reportStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Waymate API
 * Carpooling backend powering the Waymate web app.
 * OpenAPI spec version: 0.1.0
 */

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ReportStatus = {
    OPEN: "OPEN",
    INVESTIGATING: "INVESTIGATING",
    RESOLVED: "RESOLVED",
    DISMISSED: "DISMISSED",
} as const;